Date that is passed to the event table is now normalized to have the time reset to the beginning of the day.
This way the time is always standardized as all we want are unique days as keys, not NSDates with different times.  This corrects a bug where if the first date picker was not changed, it would default to an NSDate object that included the current time.  If the date was then changed and changed back to the original date, the time would be reset and the NSDate objects would not be equal, thus essentially making the original events added to the current date to be inaccessible.